chore(hypercall): Put hypercalls acting on file descriptors and paths into separate modules
