+++ src/hypercall/mod.rs
@@ -1,40 +1,26 @@
-use core::cmp;
-use std::{
-	collections::BTreeMap,
-	ffi::{CStr, CString},
-	fs, io,
-	os::{fd::IntoRawFd, unix::ffi::OsStrExt},
-	path::Path,
-	sync::Arc,
-};
+use std::io;
 
-use fstat::{fstat, stat};
-use getdents::getdents;
-use uhyve_interface::{
-	GuestPhysAddr, v1,
-	v2::{
-		self,
-		parameters::{FileType, *},
-	},
-};
+use uhyve_interface::{GuestPhysAddr, v1, v2};
 
 use crate::{
 	HypervisorResult,
-	isolation::{
-		fd::{FdData, GuestFd, UhyveFileDescriptorLayer},
-		filemap::{Directory, Node, NodeStatRef, UhyveFileMap, UhyveMapLeaf},
-	},
 	mem::MmapMemory,
 	mem_layout::MemoryLayout,
 	net::NetworkBackend,
 	vcpu::VcpuStopReason,
-	virt_to_phys,
 	vm::{KernelInfo, VmPeripherals},
 };
 
+mod by_fd;
+mod by_path;
 mod cmdval_copy;
 mod fstat;
 mod getdents;
+
+use by_fd::{close, lseek, lseek_v1, read, read_v1, write, write_v1};
+use by_path::{mkdir, open, unlink};
+use fstat::{fstat, stat};
+use getdents::getdents;
 
 /// `addr` is the address of the hypercall parameter in the guest's memory space. `data` is the
 /// parameter that was sent to that address by the guest.
@@ -284,494 +270,4 @@ fn translate_last_errno() -> Option<i32> {
 		io::Error::from_raw_os_error(errno)
 	);
 	None
-}
-
-/// Decodes a guest path given at address `path_addr` in `mem`.
-///
-/// # Safety
-///
-/// The calling convention of hypercalls ensures that the given address doesn't alias with anything mutable.
-/// The return value is only valid for the duration of the hypercall.
-unsafe fn decode_guest_path(mem: &MmapMemory, path_addr: GuestPhysAddr) -> Option<&str> {
-	let requested_path_ptr = mem.host_address(path_addr).unwrap() as *const i8;
-	unsafe { CStr::from_ptr(requested_path_ptr) }.to_str().ok()
-}
-
-fn collect_dir_entries(dir: &Directory) -> BTreeMap<Box<str>, FileType> {
-	fn host_file_type(path: &Path) -> FileType {
-		match fs::symlink_metadata(path) {
-			Ok(metadata) if metadata.is_dir() => FileType::Directory,
-			Ok(metadata) if metadata.is_file() => FileType::RegularFile,
-			Ok(metadata) if metadata.file_type().is_symlink() => FileType::SymbolicLink,
-			Ok(_) => FileType::Unknown,
-			Err(_) => FileType::RegularFile,
-		}
-	}
-
-	dir.iter()
-		.map(|(name, node)| {
-			let file_type = match node {
-				Node::Directory(_) => FileType::Directory,
-				Node::Leaf(UhyveMapLeaf::Virtual(_)) => FileType::RegularFile,
-				Node::Leaf(UhyveMapLeaf::OnHost(host_path)) => host_file_type(host_path),
-			};
-			(name.clone(), file_type)
-		})
-		.collect()
-}
-
-/// unlink deletes a name from the filesystem. This is used to handle `unlink` syscalls from the guest.
-///
-/// Note for when using Landlock: Unlinking files results in them being veiled. If a
-/// file (that existed during initialization) called `log.txt` is unlinked, attempting to
-/// open `log.txt` again will result in an error.
-fn unlink(mem: &MmapMemory, sysunlink: &mut UnlinkParams, file_map: &mut UhyveFileMap) {
-	let guest_path = if let Some(guest_path) = unsafe { decode_guest_path(mem, sysunlink.name) } {
-		guest_path
-	} else {
-		error!("The kernel requested to unlink() an non-UTF8 path: Rejecting...");
-		sysunlink.ret = -EINVAL;
-		return;
-	};
-	sysunlink.ret = match file_map.unlink(guest_path) {
-		Ok(Some(host_path)) => {
-			// We can safely unwrap here, as host_path.as_bytes will never contain internal \0 bytes
-			// As host_path_c_string is a valid CString, this implementation is presumed to be safe.
-			let host_path_c_string = CString::new(host_path.as_os_str().as_bytes()).unwrap();
-			if unsafe { libc::unlink(host_path_c_string.as_c_str().as_ptr()) } < 0 {
-				-translate_last_errno().unwrap_or(1)
-			} else {
-				0
-			}
-		}
-		Ok(None) => {
-			// Removed virtual entry
-			0
-		}
-		Err(()) => {
-			error!(
-				"The kernel requested to unlink() an unknown path ({guest_path:?}): Rejecting..."
-			);
-			-ENOENT
-		}
-	};
-}
-
-/// Handles an open syscall by opening a file on the host.
-fn open(mem: &MmapMemory, sysopen: &mut OpenParams, file_map: &mut UhyveFileMap) {
-	let guest_path = if let Some(guest_path) = unsafe { decode_guest_path(mem, sysopen.name) } {
-		guest_path
-	} else {
-		error!("The kernel requested to open() an non-UTF8 path: Rejecting...");
-		sysopen.ret = -EINVAL;
-		return;
-	};
-	let mut flags = sysopen.flags & ALLOWED_OPEN_FLAGS;
-	// See: https://lwn.net/Articles/926782/
-	// See: https://github.com/hermit-os/kernel/commit/71bc629
-	if (flags & (O_DIRECTORY | O_CREAT)) == (O_DIRECTORY | O_CREAT) {
-		error!("An open() call used O_DIRECTORY and O_CREAT at the same time. Aborting...");
-		sysopen.ret = -EINVAL;
-		return;
-	}
-
-	/// Attempts to open `host_path_c_string` with `flags` and `mode`. Inserts the fd into `fdmap`
-	/// on success and returns it, else returns the (negative) return value of the underlying `open` call.
-	fn do_open(
-		fdmap: &mut UhyveFileDescriptorLayer,
-		host_path_c_string: CString,
-		flags: i32,
-		mode: i32,
-	) -> i32 {
-		let host_fd = unsafe { libc::open(host_path_c_string.as_c_str().as_ptr(), flags, mode) };
-		if host_fd < 0 {
-			let errno = translate_last_errno().unwrap_or(1);
-			if host_fd != -1 {
-				warn!("Unexpected return value {host_fd} from open(2)");
-			}
-			-errno
-		} else if let Some(guest_fd) = fdmap.insert(FdData::Raw(host_fd)) {
-			guest_fd.0
-		} else {
-			-ENOENT
-		}
-	}
-
-	sysopen.ret = if let Some(host_node) = file_map.get_host_stat_node(guest_path, true) {
-		debug!("{guest_path:#?} found in file map.");
-		match host_node {
-			NodeStatRef::OnHost(host_path) => {
-				// We can safely unwrap here, as host_path.as_bytes will never contain internal \0 bytes
-				// As host_path_c_string is a valid CString, this implementation is presumed to be safe.
-				let host_path_c_string = CString::new(host_path.as_os_str().as_bytes()).unwrap();
-				do_open(&mut file_map.fdmap, host_path_c_string, flags, sysopen.mode)
-			}
-			NodeStatRef::VirtualFile(data) => {
-				if let Some(guest_fd) = file_map.fdmap.insert(FdData::Virtual {
-					// The following only clones a pointer, and increases an `Arc` refcount.
-					data: data.clone(),
-					offset: 0,
-				}) {
-					guest_fd.0
-				} else {
-					-ENOENT
-				}
-			}
-			NodeStatRef::VirtualDirectory(dir) => {
-				if (flags & O_DIRECTORY) != 0 {
-					if let Some(guest_fd) = file_map.fdmap.insert(FdData::MappedDirectory {
-						entries: Arc::new(collect_dir_entries(dir)),
-						offset: 0,
-					}) {
-						guest_fd.0
-					} else {
-						-ENOENT
-					}
-				} else {
-					debug!("{guest_path:#?}: Tried to open directory as file...");
-					-EINVAL
-				}
-			}
-		}
-	} else {
-		debug!("{guest_path:#?} not found in file map.");
-		if (flags & O_CREAT) == O_CREAT {
-			debug!("Attempting to open a temp file for {guest_path:#?}...");
-			// Existing files that already exist should be in the file map, not here.
-			// If a supposed attacker can predict where we open a file and its filename,
-			// this contigency, together with O_CREAT, will cause the write to fail.
-			flags |= O_EXCL;
-			#[cfg(target_os = "linux")]
-			{
-				flags |= file_map.get_io_mode_flags();
-			}
-
-			match file_map.create_temporary_file(guest_path) {
-				Some(host_path_c_string) => {
-					do_open(&mut file_map.fdmap, host_path_c_string, flags, sysopen.mode)
-				}
-				None => {
-					debug!("Returning -EINVAL for {guest_path:#?}");
-					-EINVAL
-				}
-			}
-		} else {
-			debug!("Returning -ENOENT for {guest_path:#?}");
-			-ENOENT
-		}
-	}
-}
-
-/// Handles an close syscall by closing the file on the host.
-fn close(sysclose: &mut CloseParams, file_map: &mut UhyveFileMap) {
-	let gfd = GuestFd(sysclose.fd);
-	debug!(
-		"Guest tries to close fd {gfd} from fdmap {:?}",
-		file_map.fdmap
-	);
-	sysclose.ret = if gfd.is_standard() {
-		// ignore stdio closures
-		warn!("Guest tried to close stdio fd: {gfd}");
-		0
-	} else if let Some(fddata) = file_map.fdmap.remove(gfd) {
-		if let FdData::Raw(fd) = fddata
-			&& unsafe { libc::close(fd) } < 0
-		{
-			-translate_last_errno().unwrap_or(1)
-		} else {
-			0
-		}
-	} else {
-		warn!("Guest tried to close unknown fd: {gfd}");
-		-EBADF
-	};
-}
-
-/// Handles a v1 read hypercall (for which a guest-provided guest virtual address must be
-/// converted to a guest physical address by the host).
-fn read_v1(
-	mem: &MmapMemory,
-	sysread: &mut v1::parameters::ReadParams,
-	root_pt: GuestPhysAddr,
-	file_map: &mut UhyveFileMap,
-) {
-	sysread.ret = if let Ok(guest_phys_addr) = virt_to_phys(sysread.buf, mem, root_pt) {
-		let mut tmp = v2::parameters::ReadParams {
-			fd: sysread.fd,
-			buf: guest_phys_addr,
-			len: sysread.len as u64,
-			ret: 0i64,
-		};
-		read(mem, &mut tmp, file_map);
-		tmp.ret
-			.try_into()
-			.unwrap_or_else(|ret| panic!("Unable to fit return value {} in read_v1.", ret))
-	} else {
-		warn!("Unable to convert guest virtual address into guest physical address");
-		-EFAULT as isize
-	}
-}
-
-/// Handles a read syscall on the host.
-fn read(mem: &MmapMemory, sysread: &mut v2::parameters::ReadParams, file_map: &mut UhyveFileMap) {
-	sysread.ret = if let Some(fdata) = file_map.fdmap.get_mut(GuestFd(sysread.fd.into_raw_fd())) {
-		if let Ok(host_address) = mem.host_address(sysread.buf) {
-			match fdata {
-				FdData::Raw(rfd) => {
-					let bytes_read = unsafe {
-						libc::read(
-							*rfd,
-							host_address as *mut libc::c_void,
-							sysread.len as usize,
-						)
-					};
-					if bytes_read < 0 {
-						-translate_last_errno().unwrap_or(1) as i64
-					} else {
-						bytes_read as i64
-					}
-				}
-				FdData::Virtual { data, offset } => {
-					let remaining = {
-						let pos = cmp::min(*offset, data.len() as u64);
-						&data[pos as usize..]
-					};
-					let amt = cmp::min(remaining.len() as u64, sysread.len) as usize;
-					assert!(amt <= isize::MAX as usize);
-
-					// SAFETY: the input slices can't overlap, as `host_address` is owned by the guest
-					// and `data` is owned by the host.
-					unsafe {
-						core::ptr::copy_nonoverlapping(
-							remaining.as_ptr(),
-							host_address as *mut u8,
-							amt,
-						)
-					};
-					*offset += amt as u64;
-					amt as i64
-				}
-				FdData::MappedDirectory { .. } => -EBADF as i64,
-			}
-		} else {
-			warn!("Unable to get host address for read buffer");
-			-EFAULT as i64
-		}
-	} else {
-		-EBADF as i64
-	};
-}
-
-/// Handles a v1 write hypercall (for which a guest-provided guest virtual address must be
-/// converted to a guest physical address by the host).
-fn write_v1<N: NetworkBackend>(
-	peripherals: &VmPeripherals<N>,
-	syswrite: &v1::parameters::WriteParams,
-	root_pt: GuestPhysAddr,
-	file_map: &mut UhyveFileMap,
-) -> io::Result<()> {
-	let guest_phys_addr = virt_to_phys(syswrite.buf, &peripherals.mem, root_pt).map_err(|e| {
-		io::Error::new(
-			io::ErrorKind::InvalidInput,
-			format!("invalid syswrite buffer: {e:?}"),
-		)
-	})?;
-	let mut tmp = v2::parameters::WriteParams {
-		fd: syswrite.fd,
-		buf: guest_phys_addr,
-		len: syswrite.len as u64,
-		ret: 0i64,
-	};
-	write(peripherals, &mut tmp, file_map)
-}
-
-/// Handles an write syscall on the host.
-fn write<N: NetworkBackend>(
-	peripherals: &VmPeripherals<N>,
-	syswrite: &mut v2::parameters::WriteParams,
-	file_map: &mut UhyveFileMap,
-) -> io::Result<()> {
-	let mut bytes = unsafe {
-		let guest_phys_addr = syswrite.buf;
-		peripherals
-			.mem
-			.slice_at(guest_phys_addr, syswrite.len.try_into().unwrap())
-			.map_err(|e| {
-				io::Error::new(
-					io::ErrorKind::InvalidInput,
-					format!("invalid syswrite buffer: {e:?}"),
-				)
-			})?
-	};
-
-	match file_map.fdmap.get_mut(GuestFd(syswrite.fd.into_raw_fd())) {
-		None => {
-			// We don't write anything if the file descriptor is not available,
-			// but this is OK, as writes are not necessarily guaranteed to write
-			// anything.
-			syswrite.ret = -EBADF as i64;
-			Err(io::Error::other("Bad file descriptor"))
-		}
-
-		Some(FdData::Virtual { .. }) | Some(FdData::MappedDirectory { .. }) => {
-			// virtual fds are read-only
-			syswrite.ret = -EROFS as i64;
-			Err(io::Error::new(
-				io::ErrorKind::ReadOnlyFilesystem,
-				format!(
-					"Unable to write to virtual file {}",
-					GuestFd(syswrite.fd.into_raw_fd())
-				),
-			))
-		}
-
-		// Handles to standard outputs differs to that of e.g. files.
-		Some(FdData::Raw(1 | 2)) => {
-			// Assumption: Everything is printed successfully on the host.
-			// We could assume that this will always succeed and leave it at zero, but:
-			// - having some "write" scenarios that treat a zero as an error
-			//   and some that don't is not very clean.
-			// - there is a debug_assert in the kernel that depends on this,
-			//   just in case.
-			syswrite.ret = bytes.len().try_into().unwrap();
-			peripherals.serial.output(bytes)
-		}
-
-		Some(FdData::Raw(r)) => {
-			syswrite.ret = 0;
-			while !bytes.is_empty() {
-				let step = unsafe {
-					libc::write(
-						*r,
-						&bytes[0] as *const u8 as *const libc::c_void,
-						bytes.len(),
-					)
-				};
-				if step >= 0 {
-					syswrite.ret += step as i64;
-					bytes = &bytes[step as usize..];
-				} else {
-					syswrite.ret = -translate_last_errno().unwrap_or(1) as i64;
-					return Err(io::Error::last_os_error());
-				}
-			}
-
-			Ok(())
-		}
-	}
-}
-
-/// Handles a v1 lseek syscall on the host, which has a different struct format.
-fn lseek_v1(syslseek: &mut v1::parameters::LseekParams, file_map: &mut UhyveFileMap) {
-	let mut tmp = LseekParams {
-		offset: syslseek.offset as i64,
-		whence: syslseek.whence as u32,
-		fd: syslseek.fd,
-	};
-	lseek(&mut tmp, file_map);
-	if tmp.offset < 0 {
-		tmp.offset = -1;
-	}
-	syslseek.offset = tmp
-		.offset
-		.try_into()
-		.unwrap_or_else(|ret| panic!("Unable to fit return value {} in lseek_v1.", ret));
-}
-
-/// Handles an lseek syscall on the host.
-fn lseek(syslseek: &mut LseekParams, file_map: &mut UhyveFileMap) {
-	syslseek.offset = match file_map.fdmap.get_mut(GuestFd(syslseek.fd.into_raw_fd())) {
-		Some(FdData::Raw(r)) => {
-			let ret = unsafe { libc::lseek(*r, syslseek.offset, syslseek.whence as i32) };
-			if ret < 0 {
-				-translate_last_errno().unwrap_or(1) as i64
-			} else {
-				ret
-			}
-		}
-		Some(FdData::Virtual { data, offset }) => {
-			#[forbid(unused_variables)]
-			let tmp: i64 = match syslseek.whence as i32 {
-				SEEK_SET => 0,
-				SEEK_CUR => *offset as i64,
-				SEEK_END => data.len() as i64,
-				_ => -EINVAL as i64,
-			};
-			if tmp >= 0 {
-				let tmp2 = tmp + syslseek.offset;
-				match tmp2.try_into() {
-					Ok(tmp3) => {
-						*offset = tmp3;
-						tmp2
-					}
-					_ => -EOVERFLOW as i64,
-				}
-			} else {
-				tmp
-			}
-		}
-		Some(FdData::MappedDirectory { offset, .. }) => match syslseek.whence as i32 {
-			SEEK_SET if syslseek.offset >= 0 => {
-				*offset = syslseek.offset as u64;
-				syslseek.offset
-			}
-			SEEK_CUR if syslseek.offset >= 0 => {
-				let ret = offset.saturating_add_signed(syslseek.offset);
-				*offset = ret;
-				ret as i64
-			}
-			_ => -EINVAL as i64,
-		},
-		None => {
-			warn!("lseek attempted to use an unknown file descriptor");
-			-EBADF as i64
-		}
-	};
-}
-
-/// Handles a mkdir hypercall by creating a directory on the host.
-///
-/// The guest path is resolved through the file map, so directories can only be created
-/// within mapped host directories. Unmapped paths are redirected into the sandboxed
-/// temporary directory. Virtual paths are rejected.
-/// Attempts `mkdir(host_path)` on the host, mapping the outcome to a [`MkdirResult`].
-fn host_mkdir(host_path_c: &CString) -> MkdirResult {
-	// SAFETY: `host_path_c` is a valid, null-terminated C string.
-	if unsafe { libc::mkdir(host_path_c.as_ptr(), 0o777) } < 0 {
-		MkdirResult::Error(translate_last_errno().unwrap_or(EIO))
-	} else {
-		MkdirResult::Success
-	}
-}
-
-/// Handles a mkdir hypercall by creating a directory on the host.
-///
-/// The guest path is resolved through the file map, so directories can only be created
-/// within mapped host directories. Unmapped paths are redirected into the sandboxed
-/// temporary directory. Virtual paths are rejected.
-pub(crate) fn mkdir(mem: &MmapMemory, sysmkdir: &mut MkdirParams, file_map: &mut UhyveFileMap) {
-	let Some(guest_path) = (unsafe { decode_guest_path(mem, sysmkdir.path) }) else {
-		error!("The kernel requested to mkdir() a non-UTF8 path: Rejecting...");
-		sysmkdir.ret = MkdirResult::Error(EINVAL);
-		return;
-	};
-
-	sysmkdir.ret = match file_map.get_host_path(guest_path, false) {
-		Some(UhyveMapLeaf::OnHost(host_path)) => {
-			// We can safely unwrap, as a resolved host path never contains internal NUL bytes.
-			let host_path_c = CString::new(host_path.as_os_str().as_bytes()).unwrap();
-			host_mkdir(&host_path_c)
-		}
-		Some(UhyveMapLeaf::Virtual(_)) => {
-			debug!("mkdir {guest_path:?}: target is a read-only virtual file, rejecting...");
-			MkdirResult::Error(EROFS)
-		}
-		None => {
-			debug!("mkdir {guest_path:?}: not mapped, creating a temporary directory...");
-			match file_map.create_temporary_directory(guest_path) {
-				Some(host_path_c) => host_mkdir(&host_path_c),
-				None => MkdirResult::Error(EINVAL),
-			}
-		}
-	};
 }
